requirements.txt
flask==3.1.0
python-dotenv==1.0.1
telethon==1.38.1
python-telegram-bot==20.0a1
